requirements.txt
streamlit
pandas
bertopic
sentence-transformers
umap-learn
hdbscan
plotly 
openpyxl